Pull PointerLikeTypeTraits<DeclName> up to Identifier.h
No functionality change.
